refactor: move state into shared, and rename it as store
