+++ src/app/pages/button/binary-operator-button-component.ts
@@ -1,7 +1,7 @@
 import { Component, Input } from "@angular/core";
 import { Store } from "@ngrx/store";
-import { BinaryCalcOperator } from "src/app/models/binary-operator.type";
-import { BinaryOperatorAction } from "src/app/state/calculator.actions";
+import { BinaryCalcOperator } from "src/app/shared/models/binary-operator.type";
+import { BinaryOperatorAction } from "src/app/shared/store/calculator.actions";
 import { ButtonComponent } from "./button.component";
 
 @Component({
