+++ src/app/pages/button/memory-operator-button-component.ts
@@ -1,7 +1,7 @@
 import { Component } from "@angular/core";
 import { Store } from "@ngrx/store";
-import { MemoryOperator } from "src/app/models/memory-operator.type";
-import { MemoryOperatorAction } from "src/app/state/calculator.actions";
+import { MemoryOperator } from "src/app/shared/models/memory-operator.type";
+import { MemoryOperatorAction } from "src/app/shared/store/calculator.actions";
 import { ButtonComponent } from "./button.component";
 
 @Component({
