+++ src/app/pages/button/point-button-component.ts
@@ -1,6 +1,6 @@
 import { Component } from "@angular/core";
 import { Store } from "@ngrx/store";
-import { PointAction } from "src/app/state/calculator.actions";
+import { PointAction } from "src/app/shared/store/calculator.actions";
 import { ButtonComponent } from "./button.component";
 
 @Component({
