+++ src/app/shared/models/binary-operator.type.ts
