+++ src/app/shared/models/calc-state.model.ts
