+++ src/app/shared/models/fsm-state.enum.ts
