+++ src/app/shared/models/memory-operator.type.ts
