+++ src/app/shared/models/memory-recall.type.ts
