+++ src/app/shared/models/number.type.ts
