+++ src/app/shared/models/unary-operator.type.ts
