+++ src/app/shared/store/calculator.actions.ts
