+++ src/app/shared/store/calculator.reducers.ts
