+++ src/app/shared/store/calculator.selectors.ts
