+++ src/app/shared/store/memory-operator.helper.ts
