+++ src/app/shared/store/operand.helper.ts
