+++ src/app/shared/store/reducers/binary-operator.reducer.ts
@@ -1,6 +1,6 @@
-import { BinaryCalcOperator } from "src/app/models/binary-operator.type";
-import { CalcState } from "src/app/models/calc-state.model";
-import { FsmState } from "src/app/models/fsm-state.enum";
+import { BinaryCalcOperator } from "src/app/shared/models/binary-operator.type";
+import { CalcState } from "src/app/shared/models/calc-state.model";
+import { FsmState } from "src/app/shared/models/fsm-state.enum";
 import { initialState } from "../calculator.reducers";
 import { calculate } from "../operand.helper";
