+++ src/app/shared/store/reducers/memory-operator.reducer.ts
@@ -1,6 +1,6 @@
-import { MemoryOperator } from "src/app/models/memory-operator.type";
-import { CalcState } from "src/app/models/calc-state.model";
-import { FsmState } from "src/app/models/fsm-state.enum";
+import { MemoryOperator } from "src/app/shared/models/memory-operator.type";
+import { CalcState } from "src/app/shared/models/calc-state.model";
+import { FsmState } from "src/app/shared/models/fsm-state.enum";
 import { memoryOperate } from "../memory-operator.helper";
 
 export function onMemoryOperatorReducer(calcState: CalcState, { oper }: { oper: MemoryOperator }): CalcState {
