+++ src/app/shared/store/reducers/reset.reducer.ts
@@ -1,6 +1,6 @@
-import { CalcState } from "src/app/models/calc-state.model";
-import { FsmState } from "src/app/models/fsm-state.enum";
-import { ResetType } from "src/app/models/reset-type.enum";
+import { CalcState } from "src/app/shared/models/calc-state.model";
+import { FsmState } from "src/app/shared/models/fsm-state.enum";
+import { ResetType } from "src/app/shared/models/reset-type.enum";
 import { initialState } from "../calculator.reducers";
 
 export function onResetReduer(state: CalcState, { resetType }: { resetType: ResetType }): CalcState {
