+++ src/app/shared/store/unary-operator.helper.ts
